requirements.txt
django<3
Markdown==3.1.1
django-filter==2.0.0
djangorestframework==3.9.4
psycopg2>=2.7,<3.0
django-ckeditor==5.9.0
Pillow==7.1.2
